refactor(cli): move crawl CLI to crawl_cli.py; leave run.py as import wrapper
